constants over 32b (16b fractionnal), correct orders

+++ Drivers/LL/TIMER/tim.h
@@ -14,8 +14,8 @@ void MX_TIM14_Init(void);
 void MX_TIM16_Init(void);
 void MX_TIM17_Init(void);
 
-int16_t COD_get_left();
-int32_t COD_get_right();
+int32_t COD_get_left();
+int16_t COD_get_right();
 void PWM_write(GPIO_Pin& pin, uint16_t value);
 void PWM_write_us(GPIO_Pin& pin, uint16_t width_us);
